Moved the code that processes overview DSL in configOverview and recordOverview from routes to middleware. Throw error if lable is not set in overview body DSL.
