+++ lib/formtools/overview.js
@@ -2,6 +2,28 @@
 var async = require('async'),
     cellRenderers = require('./renderers-cell');
 
+/**
+ * Set unique tabId to each tab in overview.body
+ * @param {Array} overviewBody overview body
+ */
+function setTabId (overviewBody) {
+
+    overviewBody.forEach(function (section, sectionIndex) {
+
+        if (Array.isArray(section)) {
+
+            section.forEach(function (tab, tabIndex) {
+
+                // Replace white space from tab.lable with "-", make it lowecase and add the index number at the end.
+                // sectionIndex and tabIndex number are added at the end to make the tabID unique in case when more then one tab have the same label (title)
+                tab.tabId = (tab.label).replace(/\s+/g, '-').toLowerCase() + '-' + sectionIndex + '-' + tabIndex;
+
+            });
+        }
+    });
+
+}
+
 /*
  * Replace each field name in section of overview.body with object containing field label and filed value
  * e.g ({ label: '', value: '' })
@@ -104,6 +126,11 @@ function getSections (req, res, record, model, section, cb) {
 
     async.mapSeries(section, function (obj, callback) {
 
+        // each object (each tabs) in array section of overview body DSL must have label
+        if (!obj.label) {
+            return cb(new Error('Label is required for each section in overview body.'));
+        }
+
         // recursively goes over nested section which can have repeated object structures
         return getSections(req, res, record, model, obj, callback);
 
@@ -138,6 +165,11 @@ function body (req, res, record, model, cb) {
     // Since overview.body is not a function, than it must be an Array.
     async.mapSeries(overview.body, function (section, callback) {
 
+        // each section object in overview body DSL must have label.
+        if (!Array.isArray(section) && !section.label) {
+            return cb(new Error('Label is required for each section in overview body.'));
+        }
+
         // returns an array of objects
         return getSections(req, res, record, model, section, callback);
 
@@ -146,5 +178,6 @@ function body (req, res, record, model, cb) {
 }
 
 module.exports = {
-    body: body
+    body: body,
+    setTabId: setTabId
 };
