+++ lib/formtools/templates/index.js
@@ -1,6 +1,6 @@
-var fs = require('fs'),
-    handlebars = require('handlebars');
+var handlebars = require('handlebars'),
+    path = require('path');
 
-var documents = fs.readFileSync(__dirname + '/documents.html', 'utf-8');
+var documents = path.resolve(__dirname, '/documents.html');
 
 module.exports.documents = handlebars.compile(documents);
