+++ middleware/configOverview.js
@@ -5,19 +5,25 @@ module.exports = function () {
 
 	return function (req, res, next) {
 
+        req.linz.overview = req.linz.overview || {};
+
         async.series([
 
             function (cb) {
 
                 // get doc
-                var db  = linz.mongoose.connection.db;
+                var db = linz.mongoose.connection.db;
 
                 db.collection(linz.get('configs collection name'), function (err, collection) {
 
-                    collection.findOne({ _id: req.params.config}, function (err, doc) {
+                    if (err) {
+                        return cb(err);
+                    }
 
-                        if (err) {
-                            return cb(err);
+                    collection.findOne({ _id: req.params.config}, function (findErr, doc) {
+
+                        if (findErr) {
+                            return cb(findErr);
                         }
 
                         req.linz.record = doc;
@@ -28,12 +34,27 @@ module.exports = function () {
 
                 });
 
+            },
+
+            function (cb) {
+
+                linz.formtools.overview.body(req, res, req.linz.record, req.linz.config, function (err, body) {
+
+                    if (err) {
+                        return cb(err);
+                    }
+
+                    // body could be a string of HTML content OR an array of objects
+                    req.linz.overview.body = body;
+
+                    return cb();
+
+                });
+
             }
 
-        ], function (err, results) {
-            return next(err);
-        });
+        ], next);
 
-	}
+	};
 
-}
+};
