+++ views/mixins/overviewBody.jade
@@ -34,11 +34,9 @@ mixin tabs(tabs)
 		div
 			ul.nav.nav-tabs.nav-tabs-overview
 				each tab, index in tabs
-					- var tabLink = (tab.label || '').replace(/\s+/g, '-').toLowerCase() + index;
 					li(class="#{index === 0 ? 'active' : ''}")
-						a(href='#'+'#{tabLink}', role='tab' data-toggle='tab')=tab.label
+						a(href='#'+'#{tab.tabId}', role='tab' data-toggle='tab')=tab.label
 		div.tab-content
 			each tab, index in tabs
-				- var tabId = (tab.label || '').replace(/\s+/g, '-').toLowerCase() + index;
-				div(id='#{tabId}', class="#{index === 0 ? 'tab-pane active' : 'tab-pane'}")
+				div(id='#{tab.tabId}', class="#{index === 0 ? 'tab-pane active' : 'tab-pane'}")
 					+section(tab, true)
